refactor: move jest matcher type defs to test utils
