+++ tsconfig.json
@@ -11,7 +11,7 @@
     "resolveJsonModule": true,
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
-    "typeRoots": ["./src/types", "./types", "node_modules/@types"],
+    "typeRoots": ["./src/types", "node_modules/@types"],
     "rootDirs": ["src"],
     "outDir": "dist"
   },
